[libc] Fix size_t size accross architectures

+++ libc/include/bits/i386-sysv/arch/string.h
@@ -5,5 +5,5 @@
 #endif
 
 #ifndef _SIZET
-typedef unsigned size_t;
+typedef unsigned long size_t;
 #endif
